// frontend/src/types/doctor.type.ts
export type DoctorSpecialist = {
  _id: string;
  name: string;
  slug: string;
  icon: string;
};

export type DoctorPracticeLocation = {
  clinicName: string;
  address: string;
  city: string;
};

export type DoctorSchedule = {
  day: string;
  startTime: string;
  endTime: string;
  isAvailable: boolean;
};

export type Doctor = {
  _id: string;
  user: string;
  fullName: string;
  specialist: DoctorSpecialist;
  profilePhoto: string;
  experienceYears: number;
  description: string;
  practiceLocation: DoctorPracticeLocation;
  schedule: DoctorSchedule[];
  consultationFee: number;
  ratingAverage: number;
  ratingCount: number;
  isAvailable: boolean;
  __v: number;
  createdAt: string;
  updatedAt: string;
};